feat: applay the strip payment at api and reflect that on the order response to user with isAlreadyPayed feild to show to user this order is already payed

+++ Presentation/dto/PaymentRequirmentData.cs
@@ -4,6 +4,6 @@ namespace api.Presentation.dto;
 
 public class PaymentRequirementData
 {
-    public decimal Amount{ get; set; }
+    public long Amount{ get; set; }
    [Required]  public string Currency{ get; set; }
 }
